Update dependencies
This updates the dependencies to use the latest version of Velero and
also updates the Kubernetes libraries to use the same version as the
Velero codebase.

As part of this, the external-snapshotter library has been upgraded to
use the v4 client as we are using the core Velero codebase. We are still
using the v1beta1 version of the VolumeSnapshot APIs.

Also enforce the use of the latest version of github.com/gogo/protobuf.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>